Method call passes null for nonnull parameter This method call passes a null value for a nonnull method parameter. Either the parameter is annotated as a parameter that should always be nonnull, or analysis has shown that it will always be dereferenced.

+++ atlas-loader/src/main/java/uk/ac/ebi/gxa/loader/service/AtlasMAGETABLoader.java
@@ -196,7 +196,8 @@ public class AtlasMAGETABLoader extends AtlasLoaderService {
                 throw new AtlasLoaderException(e);
             }
         } finally {
-            deleteDirectory(tempDirectory);
+            if (tempDirectory != null)
+                deleteDirectory(tempDirectory);
             try {
                 AtlasLoadCacheRegistry.getRegistry().deregisterExperiment(investigation);
             } catch (Exception e) {
